+++ package.json
@@ -67,7 +67,7 @@
     "nestia": "^9.0.3",
     "prettier": "^3.2.5",
     "prettier-plugin-prisma": "^5.0.0",
-    "prisma-markdown": "^3.0.1",
+    "prisma-markdown": "^4.0.0",
     "rimraf": "^3.0.2",
     "rollup": "^4.18.0",
     "sloc": "^0.2.1",
@@ -87,13 +87,14 @@
     "@nestjs/common": "^11.1.9",
     "@nestjs/core": "^11.1.9",
     "@nestjs/platform-express": "^11.1.9",
-    "@prisma/adapter-pg": "^6.19.0",
-    "@prisma/client": "^6.19.0",
+    "@prisma/adapter-pg": "^7.0.0",
+    "@prisma/client": "^7.0.0",
+    "@prisma/internals": "^7.0.0",
     "dotenv": "^16.3.1",
     "dotenv-expand": "^10.0.0",
     "git-last-commit": "^1.0.0",
     "jsonwebtoken": "^9.0.2",
-    "prisma": "^6.19.0",
+    "prisma": "^7.0.0",
     "serialize-error": "^4.1.0",
     "source-map-support": "^0.5.19",
     "tgrid": "^1.2.0",
