+++ src/providers/common/AttachmentFileProvider.ts
@@ -1,4 +1,4 @@
-import { Prisma } from "@prisma/client";
+import { Prisma } from "@prisma/sdk";
 import { v4 } from "uuid";
 
 import { IAttachmentFile } from "@ORGANIZATION/PROJECT-api/lib/structures/common/IAttachmentFile";
@@ -15,15 +15,15 @@ export namespace AttachmentFileProvider {
       created_at: input.created_at.toISOString(),
     });
     export const select = () =>
-      Prisma.validator<Prisma.attachment_filesFindManyArgs>()({});
+      ({}) satisfies Prisma.attachment_filesFindManyArgs;
   }
 
   export const collect = (input: IAttachmentFile.IStore) =>
-    Prisma.validator<Prisma.attachment_filesCreateInput>()({
+    ({
       id: v4(),
       name: input.name,
       extension: input.extension,
       url: input.url,
       created_at: new Date(),
-    });
+    }) satisfies Prisma.attachment_filesCreateInput;
 }
