+++ tsconfig.json
@@ -45,6 +45,7 @@
       "@ORGANIZATION/PROJECT-api/lib/*": ["./src/api/*"],
       "@ORGANIZATION/PROJECT-api/lib/": ["./src/api"],
       "@ORGANIZATION/PROJECT-api": ["./src/api"],
+      "@prisma/sdk": ["./src/prisma/client.ts"],
     }, /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
